move wav device code to its own file
