+++ vjpy/__init__.py
@@ -2,5 +2,6 @@
 
 from vjpy.data_classes import Bar, Pattern, Drumkit, Drum, NoteValue
 from vjpy.midi_device import MidiDevice
+from vjpy.wav_device import WavDevice
 from vjpy.vjpy_device import VjPyDevice
 from vjpy.data.midi.drum_patterns import patterns
