+++ vjpy/vjpy_device.py
@@ -1,19 +1,6 @@
 """vjpy backend."""
 
-import os
-import numpy as np
-from scipy.io import wavfile
-from scipy.io.wavfile import write
-from playsound import playsound
-from moviepy.editor import (
-    VideoFileClip,
-    CompositeVideoClip,
-    concatenate_videoclips,
-    clips_array,
-    vfx
-    )
-from moviepy.audio.fx.all import volumex
-from vjpy import Drumkit, Drum, NoteValue, MidiDevice
+from vjpy import Drumkit, Drum, NoteValue, MidiDevice, WavDevice
 
 
 class VjPyDevice:
@@ -35,7 +22,7 @@ class VjPyDevice:
             drumkit_sh_names=self.drumkit_sh_names
             )
 
-        self.video_device = VideoDevice()
+        # self.video_device = VideoDevice()
 
     @property
     def note_values(self):
@@ -94,63 +81,38 @@ class VjPyDevice:
             drumkit_note_names[drum.note] = drum.name
         return drumkit_note_names
 
+# from moviepy.editor import (
+#     VideoFileClip,
+#     CompositeVideoClip,
+#     concatenate_videoclips,
+#     clips_array,
+#     vfx
+#     )
+# from moviepy.audio.fx.all import volumex
 
-class WavDevice:
-    """Audio device to manipulate wav files."""
 
-    def __init__(self, drumkit_note_names, drumkit_sh_names):
-        self.sample_rate = 44100
-        self.drumkit_note_names = drumkit_note_names
-        self.drumkit_sh_names = drumkit_sh_names
+# class VideoDevice:
+#     def __init__(self):
+#         self.soundbanks_path = os.path.join('soundbanks')
 
-    def play_drum_wav_from_midi_msg(self, midi_msg):
-        """Play a drum wav file associated with a midi message."""
-        wav_name = self.drumkit_note_names[midi_msg.note]
-        playsound(f"vjpy/data/wav/drumkits/myfunkkit/{wav_name}.wav")
+#     def get_videoclip(self, video_filename):
+#         return VideoFileClip(video_filename)
 
-    def write_concatenated_wavs(self, shs):  # notes):
-        """
-        1. Take integers or shorthands, pass them as MIDI notes to mido.
+#     def get_videosubclip(self, video_clip, start=0, duration=.075):
+#         return video_clip.subclip(start, start + duration)
 
-        2. Associate them with wav file paths.
-        3. Concatenate resulting wav files.
-        4. Write and play resulting concatenated wav.
-        """
-        # msgs = [mido.Message('note_on', note=note) for note in notes]
-        wav_array = []
-        for short_hand in shs:  # for msg in msgs:
-            wav_name = f"{self.drumkit_sh_names[short_hand]}.wav"
-            # wav_name = f"{self.drumkit_note_names[msg.note]}.wav"
-            wav_path = f"vjpy/data/wav/drumkits/myfunkkit/{wav_name}"
-            _, data = wavfile.read(wav_path)
-            wav_array.append(data)
-        wav_array_c = np.concatenate(wav_array)
-        wav_array_c_name = "vjpy/data/wav/wav_examples/concat.wav"
-        write(wav_array_c_name, self.sample_rate, wav_array_c)
-        playsound(wav_array_c_name)
+#     def concatenate_subclips(self, subclips):
+#         return concatenate_videoclips(subclips)
 
-class VideoDevice:
-    def __init__(self):
-        self.soundbanks_path = os.path.join('soundbanks')
+#     def get_bars_subclips(self, bars):
+#         subclips = []
+#         for bar_ in bars:
+#             for patt in bar_:
+#                 for subclip in patt:
+#                     subclips.append(subclip)
+#         return subclips
 
-    def get_videoclip(self, video_filename):
-        return VideoFileClip(video_filename)
-
-    def get_videosubclip(self, video_clip, start=0, duration=.075):
-        return video_clip.subclip(start, start + duration)
-
-    def concatenate_subclips(self, subclips):
-        return concatenate_videoclips(subclips)
-
-    def get_bars_subclips(self, bars):
-        subclips = []
-        for bar_ in bars:
-            for patt in bar_:
-                for subclip in patt:
-                    subclips.append(subclip)
-        return subclips
-
-    def write_concatenated_subclips(self, concatenated_subclips, subclip_name):
-        concatenated_subclips.write_videofile(subclip_name)
+#     def write_concatenated_subclips(self, concatenated_subclips, subclip_name):
+#         concatenated_subclips.write_videofile(subclip_name)
 
     # todo: implement compositing
